Main services added

+++ src/main/java/ru/otus/hrapp/repository/ContractRepository.java
@@ -3,6 +3,6 @@ package ru.otus.hrapp.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.otus.hrapp.model.entity.Contract;
 
-public interface ContractRepository extends JpaRepository<Contract, Integer> {
+public interface ContractRepository extends JpaRepository<Contract, Long> {
 
 }
