+++ src/main/java/ru/otus/hrapp/repository/JobRepository.java
@@ -3,5 +3,5 @@ package ru.otus.hrapp.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.otus.hrapp.model.entity.Job;
 
-public interface JobRepository extends CrudRepository<Job, Integer> {
+public interface JobRepository extends CrudRepository<Job, Long> {
 }
