+++ src/main/java/ru/otus/hrapp/repository/LocationRepository.java
@@ -3,5 +3,5 @@ package ru.otus.hrapp.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.otus.hrapp.model.entity.Location;
 
-public interface LocationRepository extends CrudRepository<Location, Integer> {
+public interface LocationRepository extends CrudRepository<Location, Long> {
 }
